+++ src/data/defaults.js
@@ -1,6 +1,7 @@
 // eslint-disable-next-line no-unused-vars
 const defaultSettings = {
 	'context_menu': true,
+	'hide_filtered_content': false,
 	'hide_radar': true,
 	'hide_reblog_header': true,
 	'hide_recommended_blogs': true,
@@ -8,6 +9,7 @@ const defaultSettings = {
 	'hide_sponsored': true,
 	'hide_sponsored_sidebar': true,
 	'ignore_body': false,
+	'ignore_filtered_content': true,
 	'ignore_header': false,
 	'ignore_tags': false,
 	'listBlack': ['coronavirus', 'trump'],
@@ -17,5 +19,5 @@ const defaultSettings = {
 	'show_notice': true,
 	'show_tags': true,
 	'show_words': true,
-	'version': '1.6.3'
+	'version': '1.7.0'
 }; // Initialize default values.